requirements.txt
# ==============================================================================
# DEPENDENCIES - Movie Recommendation System (Simplified for Local Development)
# ==============================================================================

# Core Framework
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0,<4.0.0

# Authentication
djangorestframework-simplejwt>=5.2.0,<6.0.0

# API Utilities
django-filter>=23.0,<24.0
django-cors-headers>=4.0.0,<5.0.0
drf-spectacular>=0.26.0,<1.0.0

# HTTP Client for TMDb
requests>=2.28.0,<3.0.0

# Environment Variables
python-decouple>=3.8,<4.0

# Database
psycopg2-binary>=2.9.0,<3.0.0

# Utilities
python-dateutil>=2.8.0,<3.0.0
pytz>=2023.3

# ==============================================================================
# OPTIONAL - Uncomment when needed
# ==============================================================================

# Database (PostgreSQL - switch from SQLite later)
# psycopg2-binary>=2.9.0,<3.0.0

# Caching (Redis - for performance optimization)
# redis>=4.5.0,<5.0.0
# django-redis>=5.2.0,<6.0.0
# hiredis>=2.2.0,<3.0.0

# Background Tasks (Celery - for ML training, TMDb sync)
# celery>=5.3.0,<6.0.0
# django-celery-beat>=2.5.0,<3.0.0
# django-celery-results>=2.5.0,<3.0.0

# Machine Learning (add when implementing recommendations)
# numpy>=1.24.0,<2.0.0
# pandas>=2.0.0,<3.0.0
# scikit-learn>=1.3.0,<2.0.0
# scipy>=1.10.0,<2.0.0
# joblib>=1.3.0,<2.0.0

# AWS S3 for Model Storage
# boto3>=1.28.0,<2.0.0
# django-storages>=1.13.0,<2.0.0

# Production Server
# gunicorn>=20.1.0,<21.0.0

# Monitoring
# sentry-sdk>=1.25.0,<2.0.0
